feat: restructure IndexInfoRepository package and implement query interface for findAll
