+++ src/main/java/com/findex/controller/IndexInfoController.java
@@ -10,10 +10,10 @@ import com.findex.service.IndexInfoService;
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springdoc.core.annotations.ParameterObject;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -31,7 +31,7 @@ public class IndexInfoController {
 
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
-    public CursorPageResponse findAll(@ParameterObject IndexInfoQuery query) {
+    public CursorPageResponse findAll(@ModelAttribute IndexInfoQuery query) {
         return indexInfoService.findAll(query);
     }
