+++ src/main/java/com/findex/repository/indexinfo/IndexInfoQueryRepository.java
@@ -1,0 +1,9 @@
+package com.findex.repository.indexinfo;
+
+import com.findex.dto.indexinfo.IndexInfoQuery;
+import com.findex.dto.response.CursorPageResponse;
+
+public interface IndexInfoQueryRepository {
+
+    CursorPageResponse findAll(IndexInfoQuery q);
+}
